worked on dmf tab(show counrty)and also pdf
