+++ src/app/components/results-common/dmf/dmf.component.html
@@ -38,12 +38,12 @@
       <div class="relative flex justify-center items-center h-[1px] bg-[#DEE5F6]">
         <span
           class="routes text-center border-[1px] px-3 py-2 bg-[#0032A0] border-[#DEE5F6] rounded-md text-white text-[14px] leading-[17.07px] font-normal h-[37px] flex justify-center items-center">
-         Sub Route {{ getSubRouteNumber(i) }}
-
+          Sub Route
+          {{ (((currentChildAPIBody?.page_no ?? 1) - 1) * 25) + (i + 1) }}
         </span>
       </div>
     </div>
-    <app-dmf-card [data]="dataItem" ></app-dmf-card>
+    <app-dmf-card [data]="dataItem" [currentChildAPIBody]="currentChildAPIBody"></app-dmf-card>
   </ng-container>
 
  <app-child-paging *ngIf="currentChildAPIBody?.count > 25" [currentChildAPIBody]="currentChildAPIBody"
