+++ src/app/components/results-common/technical-routes/technical-routes.component.ts
@@ -57,8 +57,11 @@ export class TechnicalRoutesComponent {
   viewProduct: boolean = false;
 
   ngOnInit() {
-
-  // console.log('CurrentAPIBody in TechnicalRoutesComponent', this._data.ros_data[0]._id);
+    // console.log('CurrentAPIBody in TechnicalRoutesComponent', this._data);
+    // console.log('CurrentAPIBody in TechnicalRoutesComponent', this.currentChildAPIBody)
+    // console.log('CurrentAPIBody in TechnicalRoutesComponent', this.CurrentAPIBody);
+    // console.log('CurrentAPIBody in TechnicalRoutesComponent', this._itemid);
+    // console.log('CurrentAPIBody in TechnicalRoutesComponent', this.MainDataResultShow);
     this.sharedROS.rosCount$.subscribe(count => {
       if (count && count?.index === this.index) {
         this.rosCounts = count;
