+++ src/app/components/route-result/route-result.component.ts
@@ -77,6 +77,7 @@ export class RouteResultComponent {
   @Output() resetPagination: EventEmitter<any> = new EventEmitter<any>();
   @Output() handleROSChange: EventEmitter<any> = new EventEmitter<any>();
 
+   _MainDataResultShow: any;
   _currentChildAPIBody: any;
   @Input() specialCount: any;
   @Input() currentApiData: any;
@@ -84,14 +85,13 @@ export class RouteResultComponent {
   @Input() index: number | undefined;
   @Input() searchData: any;
 
-  @Input()
+ @Input()
   get dataItem() {
     return this._dataItem;
   }
   set dataItem(value: any) {
     this._dataItem = value;
   }
-
   @Input()
   get currentChildAPIBody() {
     return this._currentChildAPIBody;
@@ -111,7 +111,7 @@ export class RouteResultComponent {
   }
 
   ngOnInit() {
-    // console.log('RouteResultComponent initialized', this._dataItem);
+    console.log('RouteResultComponent initialized', this._dataItem);
     this.resultTabs = Object.values(this.utilityService.getAllTabsName());
     this.currentTabData = this.resultTabs.find((tab: any) => tab.isActive);
     this.resultTabWithKeys = this.utilityService.getAllTabsName();
@@ -120,7 +120,7 @@ export class RouteResultComponent {
     this.resultTabs.forEach(tab => {
       this.SingleDownloadCheckbox[tab.name] = false;
     });
-  
+
     //
     const Account_type = localStorage.getItem('account_type');
     const Userdata = JSON.parse(localStorage.getItem('priviledge_json') || '');
@@ -235,14 +235,11 @@ export class RouteResultComponent {
             user_id: userInfo.user_id,
             auth_token: userInfo.auth_token,
           };
-
           const privilegeKey = `user_${this.userAuth.user_id}`;
           const userPrivilege = userInfo?.privilege_json?.[privilegeKey];
-
           if (typeof window !== 'undefined' && window.localStorage && userPrivilege) {
             localStorage.setItem('priviledge_json', JSON.stringify(userPrivilege));
           }
-
           const pharmaPrivilege = userPrivilege?.['pharmvetpat-mongodb'];
           if (
             !pharmaPrivilege ||
@@ -257,12 +254,10 @@ export class RouteResultComponent {
             this.generatePDFloader = false;
             return;
           }
-
           this.userPriviledgeService.getUserTodayPriviledgesData().subscribe({
             next: (res: any) => {
               if (res?.data) {
                 todays_limit = res.data;
-
                 const remaining = pharmaPrivilege.DailyDownloadLimit - todays_limit.downloadCount;
                 if (remaining <= 0) {
                   this.handleSetLoading.emit(false);
@@ -272,16 +267,13 @@ export class RouteResultComponent {
                   this.generatePDFloader = false;
                   return;
                 }
-
                 const searchThrough = Auth_operations.getActiveformValues()?.activeForm;
                 const dataItem = this.dataItem?.[this.resultTabWithKeys.productInfo.name];
-
                 let body_main: any = {
                   report_download: true,
                   reports: [],
                   limit: pharmaPrivilege.ReportLimit,
                 };
-
                 // Push report key if tab available
                 if (this.currentTabData?.name) {
                   const tabName = this.currentTabData.name;
@@ -372,6 +364,213 @@ export class RouteResultComponent {
     });
   }
 
+  handleGeneratePDF1() {
+    this.generatePDFloader = true;
+    this.handleSetLoading.emit(true);
+    const priviledge = localStorage.getItem('priviledge_json');
+
+    const priviledge_data = JSON.parse(priviledge || '');
+    let todays_limit: any = '';
+
+    this.userPriviledgeService.getUserPriviledgesData().subscribe({
+      next: (res: any) => {
+        if (res && res?.data && res?.data?.user_info) {
+          const userInfo = res.data.user_info;
+          this.userAuth = {
+            name: userInfo.name,
+            email: userInfo.email,
+            user_id: userInfo.user_id,
+            auth_token: userInfo.auth_token,
+          };
+          let priviledge = `user_${this.userAuth?.user_id}`;
+
+          if (
+            typeof window !== 'undefined' &&
+            window.localStorage &&
+            userInfo?.priviledge_json
+          ) {
+            localStorage.setItem(
+              'priviledge_json',
+              JSON.stringify(userInfo?.privilege_json[priviledge])
+            );
+          }
+          let priviledge_data = userInfo?.privilege_json[priviledge];
+          if (
+            !priviledge_data ||
+            priviledge_data?.['pharmvetpat-mongodb']?.SplitDownload === 'false' ||
+            priviledge_data?.['pharmvetpat-mongodb']?.DownloadCount == '' ||
+            priviledge_data?.['pharmvetpat-mongodb']?.DownloadCount == 0 ||
+            priviledge_data?.['pharmvetpat-mongodb']?.DownloadCount == '0'
+          ) {
+            this.handleSetLoading.emit(false);
+            this.OpenPriviledgeModal.emit(
+              'Report download is only allowed with premium ID, please updgrade to premium account.'
+            );
+            this.generatePDFloader = false;
+            return;
+          } else {
+            this.userPriviledgeService.getUserTodayPriviledgesData().subscribe({
+              next: (res: any) => {
+                if (res && res?.data) {
+                  todays_limit = res.data;
+                  if (
+                    priviledge_data?.['pharmvetpat-mongodb']?.DailyDownloadLimit -
+                    todays_limit?.downloadCount <= 0
+                  ) {
+                    this.handleSetLoading.emit(false);
+                    this.OpenPriviledgeModal.emit(
+                      'Your daily download limit is over for this platform.'
+                    );
+                    this.generatePDFloader = false;
+                    return;
+                  }
+                  if (
+                    priviledge_data?.['pharmvetpat-mongodb']?.DailyDownloadLimit -
+                    todays_limit?.downloadCount >
+                    0
+                  ) {
+                    let id: any = '';
+                    const searchThrough = Auth_operations.getActiveformValues().activeForm;
+                    console.log('this.CurrentAPIBody---------', this._dataItem
+                    );
+                    //console.log('searchThrough', this._data._id);
+                    this.searchThrough = searchThrough;
+                    switch (searchThrough) {
+                      case searchTypes.chemicalStructure:
+                      case searchTypes.intermediateSearch:
+                        id = this._dataItem[this.resultTabWithKeys.chemicalDirectory.name][0]._id;
+                        break;
+                      case searchTypes.synthesisSearch:
+                        id = this._dataItem[this.resultTabWithKeys.technicalRoutes.name].ros_data[0]._id;;
+                        break;
+                      case searchTypes.simpleSearch:
+                      case searchTypes.advanceSearch:
+                        id = this._dataItem[this.resultTabWithKeys.productInfo.name][0]._id;
+                        break;
+                      default:
+                        id = this._dataItem[this.resultTabWithKeys.productInfo.name][0]._id;
+
+                    }
+                    console.log('No search type selected');
+                    console.log('id', id);
+                    let body_main: any = {
+                      id: id,
+                      reports: [],
+                      limit: priviledge_data?.['pharmvetpat-mongodb']?.ReportLimit,
+                    };
+                    Object.keys(this.SingleDownloadCheckbox).forEach(key => {
+                      if (this.SingleDownloadCheckbox[key]) {
+                        const mappedName = this.tabNameToReportKey[key];
+                        body_main.reports.push(mappedName || key);
+                      }
+                    });
+
+                    if (body_main?.reports?.length == 0) {
+                      alert('please select atleast 1 option');
+                      this.handleSetLoading.emit(false);
+                      this.generatePDFloader = false;
+                      return;
+                    } else {
+                      let API_MAIN = {};
+                      if (this.searchThrough === searchTypes.synthesisSearch) {
+                        API_MAIN = {
+                          api_url: this.apiUrls.technicalRoutes.reportData,
+                          body: body_main,
+                        };
+                      } else if (
+                        this.searchThrough === searchTypes.chemicalStructure ||
+                        this.searchThrough === searchTypes.intermediateSearch
+                      ) {
+                        API_MAIN = {
+                          api_url: this.apiUrls.chemicalDirectory.reportData,
+                          body: body_main,
+                        };
+                      } else {
+                        // simpleSearch or advanceSearch
+                        API_MAIN = {
+                          api_url: this.apiUrls.basicProductInfo.reportData,
+                          body: body_main,
+                        };
+                      }
+                      try {
+                        this.serviceResultTabFiltersService.getGeneratePDF(API_MAIN).subscribe({
+                          next: (resp: any) => {
+                            const blob = new Blob([resp.body!], { type: 'application/pdf' });
+                            const contentDisposition = resp.headers.get('content-disposition');
+                            const timestamp = new Date()
+                              .toISOString()
+                              .split('.')[0] // remove milliseconds
+                              .replace(/T/, '_') // replace T with _
+                              .replace(/:/g, '-'); // format time separator
+                            // Default name
+                            let filenamePrefix = 'basicProductReport';
+
+                            // Logic to update name based on search type
+                            if (
+                              this.searchThrough === searchTypes.chemicalStructure ||
+                              this.searchThrough === searchTypes.synthesisSearch ||
+                              this.searchThrough === searchTypes.intermediateSearch
+                            ) {
+                              filenamePrefix = 'technicalRouteReport';
+                            }
+
+                            // Final filename
+                            let filename = `${filenamePrefix}_${timestamp}.pdf`;
+                            //  let filename = `${this.searchThrough}Report_${timestamp}.pdf`;
+
+                            if (contentDisposition) {
+                              const match = contentDisposition.match(/filename="?([^"]+)"?/);
+                              if (match && match[1]) {
+                                filename = match[1];
+                              }
+                            }
+                            // Download logic
+                            const fileURL = URL.createObjectURL(blob);
+                            const a = document.createElement('a');
+                            a.href = fileURL;
+                            a.download = filename;
+                            document.body.appendChild(a);
+                            a.click();
+                            document.body.removeChild(a);
+                            // Clean up after successful download
+                            this.SingleDownloadCheckbox = {};
+                            // this.index = 0;
+                            // this._data = {};
+                            // this.searchThrough = '';
+
+                            this.generatePDFloader = false;
+                            this.handleSetLoading.emit(false);
+                          },
+                          error: (err: any) => {
+                            this.generatePDFloader = false;
+                            this.handleSetLoading.emit(false);
+                            console.error('Error downloading the PDF', err);
+                          },
+                        });
+                      } catch (err) {
+                        this.handleSetLoading.emit(false);
+                        this.generatePDFloader = false;
+                      }
+                    }
+                  }
+                }
+              },
+              error: (e) => {
+                this.generatePDFloader = false;
+                this.handleSetLoading.emit(false);
+                console.error('Error:', e);
+              },
+            });
+          }
+        }
+      },
+      error: (e) => {
+        this.handleSetLoading.emit(false);
+        this.generatePDFloader = false;
+        console.error('Error:', e);
+      },
+    });
+  }
 
   onChemicalDirectoryActiveTabChange(tabName: string) {
     this.activeTab = tabName;
